test(registry): de-socket registry-build-index via fetchImpl seam; relocate CLI case to integration (#664)
Threads a trailing-optional fetchImpl?: HttpClient through buildRegistryIndex
-> scanNpm/scanGithub -> fetchJson/inspectNpmPackage/inspectArchive ->
fetchWithRetry (default-to-real; production never sets it). The in-process
buildRegistryIndex unit tests now inject a fake HttpClient serving the npm/
GitHub discovery routes + archive tarballs from disk, dropping Bun.serve.

The CLI subprocess case is inherently a real process (the spawned child
cannot take an injected fetch) and is relocated to
tests/integration/registry-build-index-cli.test.ts with its real Bun.serve.

The unit file keeps spawnSync("tar") to build fixture tarballs, so it stays
in ALLOWED_SPAWN; only ALLOWED_SERVE drops it. UNIT_PURITY_BASELINE 19 -> 18.
